Format floating-point constants for readability
Remove unnecessary cast to u64 in seeded_rng

Rename send_event variable for clarity in NBackView

+++ ui/src/tasks/nback/engine.rs
@@ -374,7 +374,7 @@ impl NBackEngine {
     }
 
     fn seeded_rng(&self, mode: RunMode) -> StdRng {
-        let combined_seed = self.config.seed ^ mode.seed_tag() ^ self.run_id as u64;
+        let combined_seed = self.config.seed ^ mode.seed_tag() ^ self.run_id;
         StdRng::seed_from_u64(combined_seed)
     }
 }
